+++ lib/new_relic/agent/tracer.rb
@@ -39,6 +39,38 @@ module NewRelic
           state.current_transaction
         end
 
+        # Returns the trace_id of the current_transaction, or +nil+ if
+        # none exists.
+        #
+        # @api public
+        def current_trace_id
+          if txn = current_transaction
+            txn.trace_id
+          end
+        end
+        alias_method :trace_id, :current_trace_id
+
+        # Returns the id of the current span, or +nil+ if none exists.
+        #
+        # @api public
+        def current_span_id
+          if span = current_segment
+            span.guid
+          end
+        end
+        alias_method :span_id, :current_span_id
+
+        # Returns a boolean indicating whether the current_transaction
+        # is sampled, or +nil+ if there is no current transaction.
+        #
+        # @api public
+        def transaction_sampled?
+          if txn = current_transaction
+            txn.sampled?
+          end
+        end
+        alias_method :sampled?, :transaction_sampled?
+
         # Runs the given block of code in a transaction.
         #
         # @param [String] name reserved for New Relic internal use